+++ .circleci/notification-scripts/failure_notification.sh
@@ -6,6 +6,5 @@ set -eou pipefail
 
 SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
 
-"${SCRIPT_DIR}/install-deps.sh"
-"${SCRIPT_DIR}/slack_notification.py" "$@"
-"${SCRIPT_DIR}/failure_github_issue.py" "$@"
+"${SCRIPT_DIR}/../install-deps.sh"
+"${SCRIPT_DIR}/failure_notification.py" "$@"