+++ apps/app/src/test/scala/com/daml/network/integration/tests/SoftDomainMigrationTopologySetupIntegrationTest.scala
@@ -454,7 +454,7 @@ class SoftDomainMigrationTopologySetupIntegrationTest
     }
 
     clue("Alice validator tops up its traffic on new domain") {
-      eventually() {
+      eventually(1.minute) {
         val topupAmount =
           getTopupParameters(aliceValidatorBackend, env.environment.clock.now).topupAmount
         aliceValidatorBackend.participantClient.traffic_control