+++ apps/app/src/test/scala/com/daml/network/util/UpdateHistoryTestUtil.scala
@@ -17,8 +17,13 @@ import com.daml.network.environment.ledger.api.{
   TreeUpdate,
 }
 import com.daml.network.http.v0.definitions
+import com.daml.network.http.v0.definitions.DamlValueEncoding.members.{CompactJson, ProtobufJson}
 import com.daml.network.integration.tests.SpliceTests.TestCommon
-import com.daml.network.scan.admin.http.{LosslessScanHttpEncodings, LossyScanHttpEncodings}
+import com.daml.network.scan.admin.http.{
+  CompactJsonScanHttpEncodings,
+  ProtobufJsonScanHttpEncodings,
+  ScanHttpEncodings,
+}
 import com.daml.network.store.UpdateHistoryTestBase.{LostInScanApi, LostInStoreIngestion}
 import com.daml.network.store.{PageLimit, UpdateHistory, UpdateHistoryTestBase}
 import com.daml.ledger.javaapi.data.*
@@ -267,14 +272,26 @@ trait UpdateHistoryTestUtil extends TestCommon {
       .getUpdateHistory(
         1000,
         None,
-        true,
+        encoding = ProtobufJson,
       )
-      .map(LosslessScanHttpEncodings.httpToLapiUpdate)
+      .map(ProtobufJsonScanHttpEncodings.httpToLapiUpdate)
 
     val historyFromStoreWithoutLostData =
-      historyFromStore.map(UpdateHistoryTestBase.withoutLostData(_, mode = LostInScanApi))
+      historyFromStore
+        .map(UpdateHistoryTestBase.withoutLostData(_, mode = LostInScanApi))
+        .map(ScanHttpEncodings.makeConsistentAcrossSvs)
 
     historyFromStoreWithoutLostData should contain theSameElementsInOrderAs historyThroughApi
+
+    historyThroughApi.headOption.foreach(fromHistory => {
+      val fromPointwiseLookup =
+        ProtobufJsonScanHttpEncodings.httpToLapiUpdate(
+          scanClient.getUpdate(fromHistory.update.update.updateId, encoding = ProtobufJson)
+        )
+      fromPointwiseLookup shouldBe fromHistory
+    })
+
+    succeed
   }
 
   def compareHistoryViaScanApi(
@@ -287,6 +304,7 @@ trait UpdateHistoryTestUtil extends TestCommon {
 
     val updatesFromHistory = updateHistoryFromParticipant(ledgerBegin, dsoParty, participant)
       .map(UpdateHistoryTestBase.withoutLostData(_, mode = LostInScanApi))
+      .map(ScanHttpEncodings.makeConsistentAcrossSvs)
 
     val updatesFromScanApi = scanClient
       .getUpdateHistory(
@@ -299,9 +317,9 @@ trait UpdateHistoryTestUtil extends TestCommon {
             updatesFromHistory.head.update.recordTime.addMicros(-1L).toString,
           )
         ),
-        lossless = false,
+        encoding = CompactJson,
       )
-      .map(LossyScanHttpEncodings.httpToLapiUpdate)
+      .map(CompactJsonScanHttpEncodings.httpToLapiUpdate)
       .map(_.update)
 
     updatesFromScanApi should have length updatesFromHistory.size.toLong
@@ -321,6 +339,14 @@ trait UpdateHistoryTestUtil extends TestCommon {
         }
       }
     }
+
+    updatesFromScanApi.headOption.foreach(fromHistory => {
+      val fromPointwiseLookup =
+        CompactJsonScanHttpEncodings.httpToLapiUpdate(
+          scanClient.getUpdate(fromHistory.update.updateId, encoding = CompactJson)
+        )
+      fromPointwiseLookup.update shouldBe fromHistory
+    })
 
     succeed
   }