+++ apps/common/src/main/scala/com/daml/network/automation/TaskOutcome.scala
@@ -16,6 +16,10 @@ case class TaskSuccess(
     description: String
 ) extends TaskOutcome
 
+case class TaskFailed(
+    description: String
+) extends TaskOutcome
+
 /** Use this to denote that the task processing did not result in any change, and thus was a no-op. */
 case object TaskNoop extends TaskOutcome with PrettyPrinting {
   override def pretty: Pretty[this.type] = {