+++ canton/community/domain/src/main/scala/com/digitalasset/canton/domain/sequencing/SequencerNode.scala
@@ -649,6 +649,7 @@ class SequencerNodeBootstrap(
             syncCrypto,
             parameters.loggingConfig,
             None,
+            parameters.exitOnFatalFailures,
             loggerFactory,
             futureSupervisor,
             firstSequencerCounterServeableForSequencer, // TODO(#18401): Review this value