+++ cluster/pulumi/sv-canton/pulumiUp.ts
@@ -1,10 +1,19 @@
+import { PulumiAbortController } from '../pulumi';
 import { pulumiOptsForMigration, runForAllMigrations } from './pulumi';
+
+const abortController = new PulumiAbortController();
 
 runForAllMigrations(async (stack, migration, sv) => {
   console.log(`[migration=${migration.id}]Updating stack for ${sv}`);
-  const pulumiOpts = pulumiOptsForMigration(migration.id, sv);
-  await stack.refresh(pulumiOpts);
-  const result = await stack.up(pulumiOpts);
+  const pulumiOpts = pulumiOptsForMigration(migration.id, sv, abortController.signal);
+  await stack.refresh(pulumiOpts).catch(err => {
+    abortController.abort();
+    throw err;
+  });
+  const result = await stack.up(pulumiOpts).catch(err => {
+    abortController.abort();
+    throw err;
+  });
   console.log(`[migration=${migration.id}]Updated stack for ${sv}`);
   console.log(JSON.stringify(result.summary));
 }, false).catch(err => {