+++ docs/src/release_notes.rst
@@ -32,6 +32,10 @@ Upcoming
 * Canton
 
   * Enabled slow future logging for all components to better debug stuck nodes.
+
+  * Added a max time of 10 minutes for processing of a sequenced event before the node crashes to get restarted.
+    This mitigates cases where nodes might get stuck due to a bug and a restart recovers them.
+
 * Deployment
 
   * **Breaking** Every Helm chart with a name starting with ``cn-`` has been renamed, now
@@ -40,6 +44,11 @@ Upcoming
     replacing ``CN`` with ``SPLICE``:
 
       * ``CN_APP_LEGACY_PARTY_HINT``
+
+  * **Breaking** The script token.py was renamed to get-token.py to avoid conflicting with some
+    imported modules.
+  * ``imagePullPolicy`` is now unset by default corresponding to ``IfNotPresent``.
+    You can overwrite it using the helm value ``imagePullPolicy`` if needed.
 
 0.2.4
 -----